try added in edit func

+++ stock_controller.py
@@ -1,4 +1,6 @@
-def add(username):
+import pandas as pd
+
+def add_product(username):
     while True:
         product_category = input("Enter Product Category: ")
         try:
@@ -54,11 +56,14 @@ def add(username):
         except ValueError:
             print("Product Price Must Contain Only Numbers")
 
+
 def list_product(username):
+    df_stocks = pd.read_csv(username + "_database.txt", sep=",", engine="python", header=None,
+                            names=["Category", "Name", "Amount", "Unit", "Cost", "Price"])
 
-    df_stocks = pd.read_csv(username + "_database.txt", sep=",", engine="python", header=None, names=["Category", "Name", "Amount", "Unit", "Cost", "Price"])
+    df_stocks = df_stocks.sort_values(by=["Category"], ascending=True)
 
-    df_stocks = df_stocks.sort_values(by=["Category"])
+    df_stocks = df_stocks.reset_index(drop=True)
     df_stocks.index = df_stocks.index + 1
 
     print("\nList of Products:")
@@ -68,7 +73,10 @@ def edit_product(username):
     df_stocks = pd.read_csv(username + "_database.txt", sep=",", engine="python", header=None,
                             names=["Category", "Name", "Amount", "Unit", "Cost", "Price"])
 
-    df_stocks = df_stocks.sort_values(by=["Category"])
+    df_stocks = df_stocks.sort_values(by=["Category"], ascending=True)
+
+    df_stocks = df_stocks.reset_index(drop=True)
+
     df_stocks.index = df_stocks.index + 1
 
     list_product(username)
@@ -80,12 +88,10 @@ def edit_product(username):
     while True:
         try:
             index_to_edit = int(input("Enter The Index Of The Product Which You Would Like To Edit: ")) - 1
-
+            break
         except ValueError:
             print("Invalid Input, The Index Must Be An Integer")
             continue
-        else:
-            break
 
     print(df_stocks.iloc[index_to_edit])
 
@@ -106,14 +112,23 @@ def edit_product(username):
             while True:
                 new_category = input("Enter New Category or cancel to exit: ")
 
-                if new_category == "cancel":
+                if new_category.lower() == "cancel":
                     break
 
-                df_stocks.iloc[index_to_edit, 0] = new_category
-                df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+                try:
+                    if not new_category.replace(" ", "").isalpha():
+                        raise TypeError
 
-                print("Category Successfully Changed.")
-                break
+                    df_stocks.iloc[index_to_edit, 0] = new_category
+                    df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+
+                    print("Category Successfully Changed.")
+
+                    break
+                except TypeError:
+                    print("Product Category Must Contain Only Letters")
+
+
 
         elif choice == "2":
             while True:
@@ -122,11 +137,19 @@ def edit_product(username):
                 if new_name == "cancel":
                     break
 
-                df_stocks.iloc[index_to_edit, 1] = new_name
-                df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+                try:
+                    if not new_name.replace(" ", "").isalpha():
+                        raise TypeError
 
-                print("Name Successfully Changed.")
-                break
+                    df_stocks.iloc[index_to_edit, 1] = new_name
+                    df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+
+                    print("Name Successfully Changed.")
+
+                    break
+                except TypeError:
+                    print("Product Name Must Contain Only Letters")
+
 
         elif choice == "3":
             while True:
@@ -135,25 +158,38 @@ def edit_product(username):
                 if new_amount == "cancel":
                     break
 
-                df_stocks.iloc[index_to_edit, 2] = new_amount
-                df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+                try:
+                    float(new_amount)
 
-                print("Amount Successfully Changed.")
-                break
+                    df_stocks.iloc[index_to_edit, 2] = new_amount
+                    df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+
+                    print("Amount Successfully Changed.")
+                    break
+
+                except ValueError:
+                    print("Product Amount Must Contain Only Numbers")
 
 
         elif choice == "4":
             while True:
-                new_unit = float(input("Enter New Unit or cancel to exit: "))
+                new_unit = input("Enter New Unit or cancel to exit: ")
 
                 if new_unit == "cancel":
                     break
 
-                df_stocks.iloc[index_to_edit, 3] = new_unit
-                df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+                try:
+                    if not new_unit.isalpha():
+                        raise TypeError
 
-                print("Unit Successfully Changed.")
-                break
+                    df_stocks.iloc[index_to_edit, 3] = new_unit
+                    df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+
+                    print("Unit Successfully Changed.")
+                    break
+
+                except TypeError:
+                    print("Product Unit Must Contain Only Letters")
 
 
         elif choice == "5":
@@ -163,11 +199,17 @@ def edit_product(username):
                 if new_cost == "cancel":
                     break
 
-                df_stocks.iloc[index_to_edit, 4] = new_cost
-                df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+                try:
+                    float(new_cost)
 
-                print("Cost Successfully Changed.")
-                break
+                    df_stocks.iloc[index_to_edit, 4] = new_cost
+                    df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+
+                    print("Cost Successfully Changed.")
+                    break
+
+                except ValueError:
+                    print("Product Cost Must Contain Only Numbers")
 
 
         elif choice == "6":
@@ -177,14 +219,23 @@ def edit_product(username):
                 if new_price == "cancel":
                     break
 
-                df_stocks.iloc[index_to_edit, 5] = new_price
-                df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+                try:
+                    float(new_price)
 
-                print("Price Successfully Changed.")
-                break
+                    df_stocks.iloc[index_to_edit, 5] = new_price
+                    df_stocks.to_csv(username + "_database.txt", sep=",", index=False, header=False)
+
+                    print("Price Successfully Changed.")
+                    break
+
+                except ValueError:
+                    print("Product Price Must Contain Only Numbers")
+
 
         elif choice == "7":
             break
 
         else:
             print("Invalid Choice")
+
+edit_product("admin")
